fix statistic
