+++ src/main.js
@@ -1,8 +1,8 @@
 import {Menu} from "./components/site-menu.js";
 import {Search} from "./components/search.js";
 import {Filter} from "./components/filter.js";
+import {Statistic} from "./components/statistic.js";
 import {BoardController} from "./controllers/board.js";
-import {StatsController} from "./controllers/stats.js";
 import {totalCards, filters} from "./data.js";
 import {renderElement, MenuId} from "./util.js";
 
@@ -12,13 +12,13 @@ const mainControlElement = mainElement.querySelector(`.main__control`);
 const menu = new Menu();
 const search = new Search();
 const filter = new Filter(filters);
+const statistics = new Statistic(totalCards);
 
 renderElement(mainControlElement, menu.getElement());
 renderElement(mainElement, search.getElement());
 renderElement(mainElement, filter.getElement());
+renderElement(mainElement, statistics.getElement());
 
-const statsController = new StatsController(mainElement, totalCards);
-statsController.init();
 const boardController = new BoardController(mainElement, totalCards, filters);
 boardController.init();
 
@@ -31,11 +31,11 @@ menu.getElement().addEventListener(`change`, (evt) => {
 
   switch (evt.target.id) {
     case MenuId.TASK:
-      statsController.hide();
+      statistics.hide();
       boardController.show();
       break;
     case MenuId.STATISTIC:
-      statsController.show();
+      statistics.show();
       boardController.hide();
       break;
     case MenuId.NEW_TASK:
